Move pattern files to patterns/ namespace (Patterns::Store, ::Schema, ::Compiler, ::Exchange, ::Gossip)
