+++ lib/legion/mcp/cold_start.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require_relative 'pattern_store'
-require_relative 'pattern_exchange'
+require_relative 'patterns/store'
+require_relative 'patterns/exchange'
 
 module Legion
   module MCP
@@ -11,15 +11,15 @@ module Legion
       module_function
 
       def load_community_patterns(path: nil)
-        log.debug("[mcp][cold_start] action=load_community_patterns store_empty=#{PatternStore.empty?}")
-        return { skipped: true, reason: 'store not empty' } unless PatternStore.empty?
+        log.debug("[mcp][cold_start] action=load_community_patterns store_empty=#{Patterns::Store.empty?}")
+        return { skipped: true, reason: 'store not empty' } unless Patterns::Store.empty?
 
         path ||= configured_path
         return { skipped: true, reason: 'no path configured' } unless path
 
         log.debug("[mcp][cold_start] action=load_community_patterns path=#{path}")
 
-        PatternExchange.import_from_file(path, trust_level: :community)
+        Patterns::Exchange.import_from_file(path, trust_level: :community)
       rescue StandardError => e
         handle_exception(e, level: :error, operation: 'legion.mcp.cold_start.load_community_patterns')
         { error: e.message, imported: 0 }
