+++ lib/legion/mcp/patterns.rb
@@ -1,14 +1,13 @@
 # frozen_string_literal: true
 
 # Barrel file: requires all Tier 0 pattern routing system modules.
-# These modules live in the Legion::MCP namespace (flat) but are
-# grouped here for organizational clarity.
+# Pattern modules live in Legion::MCP::Patterns namespace.
 
-require_relative 'pattern_store'
-require_relative 'pattern_schema'
-require_relative 'pattern_compiler'
-require_relative 'pattern_exchange'
-require_relative 'pattern_gossip'
+require_relative 'patterns/store'
+require_relative 'patterns/schema'
+require_relative 'patterns/compiler'
+require_relative 'patterns/exchange'
+require_relative 'patterns/gossip'
 require_relative 'tier_router'
 require_relative 'context_guard'
 require_relative 'context_compiler'
@@ -17,3 +16,10 @@ require_relative 'state_tracker'
 require_relative 'gap_detector'
 require_relative 'cold_start'
 require_relative 'self_generate'
+
+# Backward compatibility — old flat names delegate to new namespace
+Legion::MCP::PatternStore = Legion::MCP::Patterns::Store unless Legion::MCP.const_defined?(:PatternStore, false)
+Legion::MCP::PatternSchema = Legion::MCP::Patterns::Schema unless Legion::MCP.const_defined?(:PatternSchema, false)
+Legion::MCP::PatternCompiler = Legion::MCP::Patterns::Compiler unless Legion::MCP.const_defined?(:PatternCompiler, false)
+Legion::MCP::PatternExchange = Legion::MCP::Patterns::Exchange unless Legion::MCP.const_defined?(:PatternExchange, false)
+Legion::MCP::PatternGossip = Legion::MCP::Patterns::Gossip unless Legion::MCP.const_defined?(:PatternGossip, false)
