+++ spec/legion/mcp/observer_pattern_promotion_spec.rb
@@ -2,12 +2,12 @@
 
 require 'spec_helper'
 require 'legion/mcp/observer'
-require 'legion/mcp/pattern_store'
+require 'legion/mcp/patterns/store'
 
 RSpec.describe 'Observer pattern promotion' do
   before do
     Legion::MCP::Observer.reset!
-    Legion::MCP::PatternStore.reset!
+    Legion::MCP::Patterns::Store.reset!
   end
 
   it 'promotes a candidate after 3 identical intent+tool observations' do
@@ -20,7 +20,7 @@ RSpec.describe 'Observer pattern promotion' do
     end
 
     hash = Digest::SHA256.hexdigest('check deploy status')
-    pattern = Legion::MCP::PatternStore.lookup(hash)
+    pattern = Legion::MCP::Patterns::Store.lookup(hash)
     expect(pattern).not_to be_nil
     expect(pattern[:confidence]).to eq(0.5)
     expect(pattern[:tool_chain]).to eq(['legion.get_status'])
@@ -36,7 +36,7 @@ RSpec.describe 'Observer pattern promotion' do
     end
 
     hash = Digest::SHA256.hexdigest('check deploy status')
-    expect(Legion::MCP::PatternStore.lookup(hash)).to be_nil
+    expect(Legion::MCP::Patterns::Store.lookup(hash)).to be_nil
   end
 
   it 'does not promote failed intents' do
@@ -49,6 +49,6 @@ RSpec.describe 'Observer pattern promotion' do
     end
 
     hash = Digest::SHA256.hexdigest('failing action')
-    expect(Legion::MCP::PatternStore.lookup(hash)).to be_nil
+    expect(Legion::MCP::Patterns::Store.lookup(hash)).to be_nil
   end
 end
