+++ spec/legion/mcp/tier0_accounting_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 require 'digest'
-require 'legion/mcp/pattern_store'
+require 'legion/mcp/patterns/store'
 require 'legion/mcp/context_guard'
 require 'legion/mcp/tier_router'
 require 'legion/mcp/observer'
@@ -11,11 +11,11 @@ RSpec.describe 'Tier 0 pattern learning and promotion accounting' do
   let(:logger) { spy('logger') }
 
   before do
-    Legion::MCP::PatternStore.reset!
+    Legion::MCP::Patterns::Store.reset!
     Legion::MCP::ContextGuard.reset!
     Legion::MCP::Observer.reset!
     allow(Legion::MCP::TierRouter).to receive(:log).and_return(logger)
-    allow(Legion::MCP::PatternStore).to receive(:log).and_return(logger)
+    allow(Legion::MCP::Patterns::Store).to receive(:log).and_return(logger)
     allow(Legion::MCP::Observer).to receive(:log).and_return(logger)
   end
 
@@ -27,7 +27,7 @@ RSpec.describe 'Tier 0 pattern learning and promotion accounting' do
     let(:incoming_hash) { Digest::SHA256.hexdigest('similar but different intent') }
 
     before do
-      Legion::MCP::PatternStore.store(
+      Legion::MCP::Patterns::Store.store(
         intent_hash:   stored_hash,
         intent_text:   'check deploy status',
         intent_vector: [0.1, 0.9, 0.3, 0.5],
@@ -43,34 +43,34 @@ RSpec.describe 'Tier 0 pattern learning and promotion accounting' do
     it 'records hit against the matched pattern hash on success' do
       # Simulate semantic lookup returning the stored pattern for a different intent
       allow(Legion::MCP::TierRouter).to receive(:try_semantic_lookup).and_return(
-        Legion::MCP::PatternStore.lookup(stored_hash)
+        Legion::MCP::Patterns::Store.lookup(stored_hash)
       )
       allow(Legion::MCP::TierRouter).to receive(:execute_tool_chain)
         .and_return([{ status: 'running' }])
 
       Legion::MCP::TierRouter.route(intent: 'similar but different intent')
 
-      stored = Legion::MCP::PatternStore.lookup(stored_hash)
+      stored = Legion::MCP::Patterns::Store.lookup(stored_hash)
       expect(stored[:hit_count]).to eq(6)
     end
 
     it 'records miss against the matched pattern hash on tool chain failure' do
       allow(Legion::MCP::TierRouter).to receive(:try_semantic_lookup).and_return(
-        Legion::MCP::PatternStore.lookup(stored_hash)
+        Legion::MCP::Patterns::Store.lookup(stored_hash)
       )
       allow(Legion::MCP::TierRouter).to receive(:execute_tool_chain)
         .and_raise(StandardError, 'tool failed')
 
       Legion::MCP::TierRouter.route(intent: 'similar but different intent')
 
-      stored = Legion::MCP::PatternStore.lookup(stored_hash)
+      stored = Legion::MCP::Patterns::Store.lookup(stored_hash)
       expect(stored[:miss_count]).to eq(1)
       expect(stored[:confidence]).to be < 0.9
     end
 
     it 'does not record against the incoming intent hash' do
       allow(Legion::MCP::TierRouter).to receive(:try_semantic_lookup).and_return(
-        Legion::MCP::PatternStore.lookup(stored_hash)
+        Legion::MCP::Patterns::Store.lookup(stored_hash)
       )
       allow(Legion::MCP::TierRouter).to receive(:execute_tool_chain)
         .and_return([{ status: 'running' }])
@@ -78,7 +78,7 @@ RSpec.describe 'Tier 0 pattern learning and promotion accounting' do
       Legion::MCP::TierRouter.route(intent: 'similar but different intent')
 
       # The incoming intent hash should not have a pattern stored
-      expect(Legion::MCP::PatternStore.lookup(incoming_hash)).to be_nil
+      expect(Legion::MCP::Patterns::Store.lookup(incoming_hash)).to be_nil
     end
   end
 
@@ -107,7 +107,7 @@ RSpec.describe 'Tier 0 pattern learning and promotion accounting' do
 
       # Neither should be promoted yet (threshold is 3)
       hash = Digest::SHA256.hexdigest('check status')
-      expect(Legion::MCP::PatternStore.lookup(hash)).to be_nil
+      expect(Legion::MCP::Patterns::Store.lookup(hash)).to be_nil
     end
 
     it 'promotes only the tool that reaches threshold independently' do
@@ -121,7 +121,7 @@ RSpec.describe 'Tier 0 pattern learning and promotion accounting' do
       end
 
       hash = Digest::SHA256.hexdigest('check status')
-      pattern = Legion::MCP::PatternStore.lookup(hash)
+      pattern = Legion::MCP::Patterns::Store.lookup(hash)
       expect(pattern).not_to be_nil
       expect(pattern[:tool_chain]).to eq(['legion.tool_a'])
     end
@@ -144,7 +144,7 @@ RSpec.describe 'Tier 0 pattern learning and promotion accounting' do
       end
 
       hash = Digest::SHA256.hexdigest('run diagnostics')
-      expect(Legion::MCP::PatternStore.lookup(hash)).to be_nil
+      expect(Legion::MCP::Patterns::Store.lookup(hash)).to be_nil
     end
   end
