+++ spec/legion/mcp/tier0_integration_spec.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
-require 'legion/mcp/pattern_store'
+require 'legion/mcp/patterns/store'
 require 'legion/mcp/context_guard'
 require 'legion/mcp/tier_router'
 require 'legion/mcp/observer'
 
 RSpec.describe 'Tier 0 Integration' do
   before do
-    Legion::MCP::PatternStore.reset!
+    Legion::MCP::Patterns::Store.reset!
     Legion::MCP::ContextGuard.reset!
     Legion::MCP::Observer.reset!
   end
@@ -37,14 +37,14 @@ RSpec.describe 'Tier 0 Integration' do
 
       # Pattern should now exist
       hash = Digest::SHA256.hexdigest('audit the tools')
-      pattern = Legion::MCP::PatternStore.lookup(hash)
+      pattern = Legion::MCP::Patterns::Store.lookup(hash)
       expect(pattern).not_to be_nil
       expect(pattern[:confidence]).to eq(0.5)
 
       # Phase 2: Simulate confidence growth (manual for testing)
       pattern[:confidence] = 0.9
       pattern[:last_hit_at] = Time.now
-      Legion::MCP::PatternStore.store(pattern)
+      Legion::MCP::Patterns::Store.store(pattern)
 
       # Phase 3: Route should return Tier 0
       result = Legion::MCP::TierRouter.route(intent: 'audit the tools')
@@ -55,7 +55,7 @@ RSpec.describe 'Tier 0 Integration' do
 
   describe 'degradation: no cache, no local' do
     it 'works with L0 only (in-memory)' do
-      Legion::MCP::PatternStore.store(
+      Legion::MCP::Patterns::Store.store(
         intent_hash: Digest::SHA256.hexdigest('test intent'),
         intent_text: 'test intent',
         tool_chain:  ['legion.tool_audit'],
@@ -77,7 +77,7 @@ RSpec.describe 'Tier 0 Integration' do
   describe 'escalation on failure' do
     it 'demotes pattern after tool chain failure' do
       hash = Digest::SHA256.hexdigest('failing intent')
-      Legion::MCP::PatternStore.store(
+      Legion::MCP::Patterns::Store.store(
         intent_hash: hash, intent_text: 'failing intent',
         tool_chain:  ['legion.nonexistent_tool'],
         confidence:  0.85, hit_count: 10, miss_count: 0,
@@ -89,7 +89,7 @@ RSpec.describe 'Tier 0 Integration' do
       result = Legion::MCP::TierRouter.route(intent: 'failing intent')
       expect(result[:tier]).to eq(1)
 
-      pattern = Legion::MCP::PatternStore.lookup(hash)
+      pattern = Legion::MCP::Patterns::Store.lookup(hash)
       expect(pattern[:miss_count]).to eq(1)
       expect(pattern[:confidence]).to be < 0.85
     end
@@ -98,7 +98,7 @@ RSpec.describe 'Tier 0 Integration' do
   describe 'semantic matching' do
     it 'matches similar intents via cosine similarity' do
       vec = [0.1, 0.9, 0.3, 0.5]
-      Legion::MCP::PatternStore.store(
+      Legion::MCP::Patterns::Store.store(
         intent_hash: 'exact_hash', intent_text: 'check deploy status',
         intent_vector: vec, tool_chain: ['legion.get_status'],
         confidence: 0.9, hit_count: 5, miss_count: 0,
@@ -106,7 +106,7 @@ RSpec.describe 'Tier 0 Integration' do
       )
 
       similar_vec = [0.11, 0.89, 0.31, 0.49]
-      result = Legion::MCP::PatternStore.lookup_semantic(similar_vec, threshold: 0.99)
+      result = Legion::MCP::Patterns::Store.lookup_semantic(similar_vec, threshold: 0.99)
       expect(result).not_to be_nil
       expect(result[:intent_text]).to eq('check deploy status')
     end
