+++ spec/legion/mcp/tools/do_action_spec.rb
@@ -22,7 +22,7 @@ RSpec.describe Legion::MCP::Tools::DoAction do
     before do
       allow(Legion::MCP::Tools::DoAction).to receive(:log).and_return(logger)
       allow(Legion::MCP::TierRouter).to receive(:log).and_return(logger) if defined?(Legion::MCP::TierRouter)
-      allow(Legion::MCP::PatternStore).to receive(:log).and_return(logger) if defined?(Legion::MCP::PatternStore)
+      allow(Legion::MCP::Patterns::Store).to receive(:log).and_return(logger) if defined?(Legion::MCP::Patterns::Store)
     end
 
     context 'when no matching tool is found' do
@@ -203,7 +203,7 @@ RSpec.describe Legion::MCP::Tools::DoAction do
     describe 'Tier 0 routing' do
       before do
         require 'legion/mcp/tier_router'
-        Legion::MCP::PatternStore.reset!
+        Legion::MCP::Patterns::Store.reset!
         Legion::MCP::ContextGuard.reset!
       end
 
@@ -240,7 +240,7 @@ RSpec.describe Legion::MCP::Tools::DoAction do
 
       before do
         require 'legion/mcp/tier_router'
-        Legion::MCP::PatternStore.reset!
+        Legion::MCP::Patterns::Store.reset!
         Legion::MCP::ContextGuard.reset!
       end
 
@@ -283,7 +283,7 @@ RSpec.describe Legion::MCP::Tools::DoAction do
 
       before do
         require 'legion/mcp/tier_router'
-        Legion::MCP::PatternStore.reset!
+        Legion::MCP::Patterns::Store.reset!
         Legion::MCP::ContextGuard.reset!
       end
